app/requirements.txt
python-telegram-bot==13.15
opencv-python-headless==4.10.0.82
pytesseract==0.3.10
easyocr==1.7.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
numpy>=1.23,<2.0